// src/web/src/app/api/bff/models/institution-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface InstitutionSummary {
  additionalData?: ({
} | null);
  code?: (string | null);
  countryCode?: (string | null);
  did?: (string | null);
  displayName?: (string | null);
  id?: (string | null);
  isActive?: (boolean | null);
  issuerWalletAddress?: (string | null);
  legalName?: (string | null);
  registeredAt?: (string | null);
  websiteUrl?: (string | null);
}